requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.32.0
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.23
pydantic==2.9.2
pydantic[email]==2.9.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
httpx==0.27.2
python-dotenv==1.0.1
PyJWT==2.9.0
bcrypt==4.2.0